# backend/internal/platform/http/openapi.yaml
openapi: 3.1.0
info:
  title: Second Brain API
  version: 1.0.0
  description: REST API for a local-first AI memory system.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
tags:
  - name: Auth
  - name: Memories
  - name: Tasks
  - name: Reminders
  - name: Search
  - name: Settings
  - name: Integrations
paths:
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionEnvelope'
  /auth/logout:
    post:
      tags: [Auth]
      security: []
      summary: Revoke refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Logged out
  /memories:
    get:
      tags: [Memories]
      summary: List memories
      parameters:
        - $ref: '#/components/parameters/Query'
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Memory list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryListEnvelope'
    post:
      tags: [Memories]
      summary: Save a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemoryRequest'
            examples:
              siri:
                value:
                  content: Remember that Kafka rebalance work needs a diagram.
                  source: siri
      responses:
        '201':
          description: Memory saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryEnvelope'
  /memories/{id}:
    get:
      tags: [Memories]
      summary: Get a memory
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Memory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemoryEnvelope'
    patch:
      tags: [Memories]
      summary: Update a memory
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemoryRequest'
      responses:
        '200':
          description: Memory updated
    delete:
      tags: [Memories]
      summary: Delete a memory
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Deleted
  /memories/{id}/similar:
    get:
      tags: [Memories, Search]
      summary: Find similar memories using pgvector
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Similar results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchEnvelope'
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      parameters:
        - $ref: '#/components/parameters/Query'
        - name: status
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListEnvelope'
    post:
      tags: [Tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              voice:
                value:
                  text: Create task study Netpoll before Sunday
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskEnvelope'
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get a task
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Task
    patch:
      tags: [Tasks]
      summary: Update a task
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
    delete:
      tags: [Tasks]
      summary: Delete a task
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Deleted
  /tasks/{id}/complete:
    post:
      tags: [Tasks]
      summary: Mark task complete
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Completed task
  /reminders:
    get:
      tags: [Reminders]
      summary: List reminders
      parameters:
        - name: include_sent
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Reminder list
    post:
      tags: [Reminders]
      summary: Create reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
      responses:
        '201':
          description: Reminder created
  /search:
    get:
      tags: [Search]
      summary: Global semantic search
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchEnvelope'
  /settings:
    get:
      tags: [Settings]
      summary: Get user settings
      responses:
        '200':
          description: Settings
    put:
      tags: [Settings]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSettingsRequest'
      responses:
        '200':
          description: Settings updated
  /settings/telegram:
    put:
      tags: [Settings, Integrations]
      summary: Configure Telegram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramConfigRequest'
      responses:
        '200':
          description: Telegram config saved
  /settings/linear:
    put:
      tags: [Settings, Integrations]
      summary: Configure Linear
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinearConfigRequest'
      responses:
        '200':
          description: Linear config saved
  /telegram/test:
    post:
      tags: [Integrations]
      summary: Send Telegram test notification
      responses:
        '200':
          description: Test sent
  /linear/issues:
    get:
      tags: [Integrations]
      summary: List active Linear issues
      responses:
        '200':
          description: Linear issues
    post:
      tags: [Integrations]
      summary: Create Linear issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLinearIssueRequest'
      responses:
        '201':
          description: Linear issue created
  /linear/issues/{id}:
    patch:
      tags: [Integrations]
      summary: Update Linear issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLinearIssueRequest'
      responses:
        '200':
          description: Linear issue updated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Query:
      name: q
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    RegisterRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 1
          maxLength: 160
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    SessionEnvelope:
      type: object
      properties:
        success:
          const: true
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            user:
              $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    Memory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        summary:
          type: string
        category:
          enum: [Work, Learning, Personal, Projects, Meetings, Ideas]
        source:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        created_at:
          type: string
          format: date-time
    CreateMemoryRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        source:
          enum: [manual, siri, widget, telegram, import]
    UpdateMemoryRequest:
      type: object
      properties:
        content:
          type: string
        summary:
          type: string
        category:
          enum: [Work, Learning, Personal, Projects, Meetings, Ideas]
        tags:
          type: array
          items:
            type: string
    MemoryEnvelope:
      type: object
      properties:
        success:
          const: true
        data:
          $ref: '#/components/schemas/Memory'
    MemoryListEnvelope:
      type: object
      properties:
        success:
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Memory'
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        notes:
          type: string
        priority:
          enum: [low, medium, high, urgent]
        status:
          enum: [inbox, today, upcoming, overdue, completed]
        due_at:
          type: [string, 'null']
          format: date-time
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    CreateTaskRequest:
      type: object
      properties:
        text:
          type: string
        title:
          type: string
        notes:
          type: string
        due_at:
          type: [string, 'null']
          format: date-time
        priority:
          enum: [low, medium, high, urgent]
        tags:
          type: array
          items:
            type: string
    UpdateTaskRequest:
      allOf:
        - $ref: '#/components/schemas/CreateTaskRequest'
        - type: object
          properties:
            status:
              enum: [inbox, today, upcoming, overdue, completed]
    TaskEnvelope:
      type: object
      properties:
        success:
          const: true
        data:
          $ref: '#/components/schemas/Task'
    TaskListEnvelope:
      type: object
      properties:
        success:
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    CreateReminderRequest:
      type: object
      required: [fire_at, title]
      properties:
        task_id:
          type: [string, 'null']
          format: uuid
        memory_id:
          type: [string, 'null']
          format: uuid
        fire_at:
          type: string
          format: date-time
        title:
          type: string
        body:
          type: string
    SearchResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          enum: [memory, task]
        title:
          type: string
        snippet:
          type: string
        score:
          type: number
          format: double
        category:
          type: string
    SearchEnvelope:
      type: object
      properties:
        success:
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
    UpdateSettingsRequest:
      type: object
      properties:
        gemini_enabled:
          type: boolean
        telegram_enabled:
          type: boolean
        linear_enabled:
          type: boolean
        notifications_enabled:
          type: boolean
        siri_enabled:
          type: boolean
    TelegramConfigRequest:
      type: object
      required: [bot_token, chat_id]
      properties:
        bot_token:
          type: string
        chat_id:
          type: string
        enabled:
          type: boolean
    LinearConfigRequest:
      type: object
      required: [api_key, team_id]
      properties:
        api_key:
          type: string
        team_id:
          type: string
        enabled:
          type: boolean
    CreateLinearIssueRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
    UpdateLinearIssueRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        state_id:
          type: string